FileArgumentParser: remove HSC-specific tweaks (rerun)
It seems that the FileArgumentParser.parse_args derived from the HSC
ArgumentParser, rather than the LSST version, which made it incompatible
with the LSST version.

This reveals the need for a finer granularity in ArgumentParser.parse_args
in pipe_base, as it's been necessary to cut/paste a lot of code.